checklist_project_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg[binary]==3.2.12
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
email-validator==2.2.0
